fixed checkbox
