--- app/src/main/java/com/example/movietheatre/core/domain/util/error/LoginError.kt
@@ -1,8 +1,0 @@
-package com.example.movietheatre.core.domain.util.error
-
-
-sealed class LoginError : RootError {
-    object NoInternetError : LoginError()
-    object UserNotFound : LoginError()
-    data object UnknownError : LoginError()
-}
