--- app/src/main/java/com/example/movietheatre/core/domain/util/error/RepeatPasswordError.kt
@@ -1,6 +1,0 @@
-package com.example.movietheatre.core.domain.util.error
-
-enum class RepeatPasswordError : RootError {
-    NO_MATCH,
-    BLANK_FIELD,
-}
