+++ app/src/main/java/com/example/movietheatre/core/domain/util/error/RootError.kt
@@ -1,3 +1,3 @@
 package com.example.movietheatre.core.domain.util.error
 
-sealed interface RootError
+interface RootError
