+++ app/src/main/java/com/example/movietheatre/core/presentation/extension/ViewExtension.kt
@@ -1,9 +1,16 @@
 package com.example.movietheatre.core.presentation.extension
 
+import android.content.Context
 import android.view.View
+import android.view.inputmethod.InputMethodManager
 import com.google.android.material.snackbar.Snackbar
 
 
 fun View.showSnackBar(message: String) {
     Snackbar.make(this, message, Snackbar.LENGTH_SHORT).show()
+}
+
+fun View.hideKeyboard() {
+    val imm = context.getSystemService(Context.INPUT_METHOD_SERVICE) as InputMethodManager
+    imm.hideSoftInputFromWindow(windowToken, 0)
 }
