+++ app/src/main/java/com/example/movietheatre/feature_login/data/repository/LoginRepositoryImpl.kt
@@ -2,10 +2,11 @@ package com.example.movietheatre.feature_login.data.repository
 
 import android.util.Log
 import com.example.movietheatre.core.domain.util.Resource
-import com.example.movietheatre.core.domain.util.error.LoginError
 import com.example.movietheatre.feature_login.domain.repository.LoginRepository
+import com.example.movietheatre.feature_login.domain.util.LoginError
 import com.google.firebase.FirebaseNetworkException
 import com.google.firebase.auth.FirebaseAuth
+import com.google.firebase.auth.FirebaseAuthInvalidCredentialsException
 import com.google.firebase.auth.FirebaseAuthInvalidUserException
 import javax.inject.Inject
 import kotlin.coroutines.resume
@@ -30,6 +31,7 @@ class LoginRepositoryImpl @Inject constructor(private val firebaseAuth: Firebase
                     val error = when (exception) {
                         is FirebaseAuthInvalidUserException -> LoginError.UserNotFound
                         is FirebaseNetworkException -> LoginError.NoInternetError
+                        is FirebaseAuthInvalidCredentialsException -> LoginError.InvalidCredentials
                         else -> LoginError.UnknownError
                     }
                     cont.resume(Resource.Error(error))
