+++ app/src/main/java/com/example/movietheatre/feature_login/domain/use_case/LoginUseCase.kt
@@ -1,8 +1,8 @@
 package com.example.movietheatre.feature_login.domain.use_case
 
 import com.example.movietheatre.core.domain.util.Resource
-import com.example.movietheatre.core.domain.util.error.LoginError
 import com.example.movietheatre.feature_login.domain.repository.LoginRepository
+import com.example.movietheatre.feature_login.domain.util.LoginError
 import javax.inject.Inject
 
 class LoginUseCase @Inject constructor(private val loginRepository: LoginRepository) {
