+++ app/src/main/java/com/example/movietheatre/feature_login/presentation/screen/LoginSideEffect.kt
@@ -1,8 +1,7 @@
 package com.example.movietheatre.feature_login.presentation.screen
 
-import com.example.movietheatre.core.domain.util.error.LoginError
 
 sealed interface LoginSideEffect {
     object SuccessFullLogin : LoginSideEffect
-    data class ShowSnackBar(val message: LoginError) : LoginSideEffect
+    data class ShowSnackBar(val message: Int) : LoginSideEffect
 }
