+++ app/src/main/java/com/example/movietheatre/feature_register/domain/repository/RegisterRepository.kt
@@ -1,9 +1,9 @@
 package com.example.movietheatre.feature_register.domain.repository
 
 import com.example.movietheatre.core.domain.util.Resource
-import com.example.movietheatre.core.domain.util.error.RegisterError
+import com.example.movietheatre.feature_register.domain.util.RegisterError
 
 interface RegisterRepository {
 
-    suspend fun register(email:String,password:String):Resource<Unit,RegisterError>
+    suspend fun register(email:String,password:String):Resource<Unit, RegisterError>
 }
