+++ app/src/main/java/com/example/movietheatre/feature_register/domain/util/RegisterError.kt
@@ -1,4 +1,6 @@
-package com.example.movietheatre.core.domain.util.error
+package com.example.movietheatre.feature_register.domain.util
+
+import com.example.movietheatre.core.domain.util.error.RootError
 
 sealed class RegisterError : RootError {
     object NoInternetError : RegisterError()
