+++ app/src/main/res/color/auth_buttons_text_colors.xml
@@ -1,0 +1,5 @@
+<?xml version="1.0" encoding="utf-8"?>
+<selector xmlns:android="http://schemas.android.com/apk/res/android">
+    <item android:color="@color/black" android:state_enabled="true" />
+    <item android:color="@color/gray" android:state_enabled="false" />
+</selector>
