+++ app/src/main/res/drawable/background_auth_button_disabled.xml
@@ -1,0 +1,6 @@
+<?xml version="1.0" encoding="utf-8"?>
+<shape xmlns:android="http://schemas.android.com/apk/res/android"
+    android:shape="rectangle">
+    <solid android:color="@color/dark_white" />
+    <corners android:radius="8dp" />
+</shape>
