+++ app/src/main/res/drawable/ic_checkbox_checked.xml
@@ -1,0 +1,13 @@
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    android:width="36dp"
+    android:height="36dp"
+    android:viewportWidth="28.8"
+    android:viewportHeight="28.8">
+  <path
+      android:pathData="M10.4,14.4L13.4,17.4L18.4,11.4M6.4,19.2L6.4,9.6C6.4,8.48 6.4,7.92 6.618,7.492C6.81,7.115 7.115,6.81 7.492,6.618C7.92,6.4 8.48,6.4 9.6,6.4L19.2,6.4C20.32,6.4 20.88,6.4 21.307,6.618C21.684,6.81 21.99,7.115 22.182,7.492C22.4,7.919 22.4,8.479 22.4,9.597L22.4,19.204C22.4,20.321 22.4,20.88 22.182,21.308C21.99,21.684 21.684,21.99 21.307,22.182C20.88,22.4 20.321,22.4 19.203,22.4L9.597,22.4C8.479,22.4 7.919,22.4 7.492,22.182C7.115,21.99 6.81,21.684 6.618,21.308C6.4,20.88 6.4,20.32 6.4,19.2Z"
+      android:strokeLineJoin="round"
+      android:strokeWidth="0.8800000000000001"
+      android:fillColor="#00000000"
+      android:strokeColor="#ffffff"
+      android:strokeLineCap="round"/>
+</vector>
