+++ app/src/main/res/drawable/ic_checkbox_unchecked.xml
@@ -1,0 +1,13 @@
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    android:width="36dp"
+    android:height="36dp"
+    android:viewportWidth="28.8"
+    android:viewportHeight="28.8">
+    <path
+        android:fillColor="#00000000"
+        android:pathData="M6.4,9.6L6.4,19.2C6.4,20.32 6.4,20.88 6.618,21.308C6.81,21.684 7.115,21.99 7.492,22.182C7.919,22.4 8.479,22.4 9.597,22.4L19.203,22.4C20.321,22.4 20.88,22.4 21.307,22.182C21.684,21.99 21.99,21.684 22.182,21.308C22.4,20.88 22.4,20.321 22.4,19.204L22.4,9.597C22.4,8.479 22.4,7.919 22.182,7.492C21.99,7.115 21.684,6.81 21.307,6.618C20.88,6.4 20.32,6.4 19.2,6.4L9.6,6.4C8.48,6.4 7.92,6.4 7.492,6.618C7.115,6.81 6.81,7.115 6.618,7.492C6.4,7.92 6.4,8.48 6.4,9.6Z"
+        android:strokeWidth="0.8800000000000001"
+        android:strokeColor="#ffffff"
+        android:strokeLineCap="round"
+        android:strokeLineJoin="round" />
+</vector>
