+++ app/src/main/res/drawable/select_checkbox.xml
@@ -1,0 +1,6 @@
+<?xml version="1.0" encoding="utf-8"?>
+<selector xmlns:android="http://schemas.android.com/apk/res/android">
+    <item android:drawable="@drawable/ic_checkbox_checked" android:state_checked="true" />
+    <item android:drawable="@drawable/ic_checkbox_unchecked" android:state_checked="false" />
+
+</selector>
