+++ app/src/main/res/layout/fragment_login.xml
@@ -5,7 +5,6 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/blue"
-    android:padding="20dp"
     tools:context=".feature_login.presentation.screen.LoginFragment">
 
 
@@ -13,10 +12,10 @@
         android:id="@+id/txtLoginHeader"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
-        android:layout_marginTop="2dp"
+        android:layout_marginTop="8dp"
         android:text="@string/login"
         android:textColor="@color/white"
-        android:textSize="24sp"
+        android:textSize="36sp"
         android:textStyle="bold"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintStart_toStartOf="parent"
@@ -26,11 +25,14 @@
         android:id="@+id/etEmail"
         android:layout_width="0dp"
         android:layout_height="wrap_content"
+        android:layout_marginStart="20dp"
         android:layout_marginTop="24dp"
+        android:layout_marginEnd="20dp"
         android:background="@drawable/background_form_rounded"
         android:drawableEnd="@drawable/ic_email"
         android:drawablePadding="12dp"
         android:hint="@string/enter_email"
+        android:inputType="textEmailAddress"
         android:maxLines="1"
         android:paddingStart="12dp"
         android:paddingTop="12dp"
@@ -69,6 +71,8 @@
             android:id="@+id/etPassword"
             android:layout_width="match_parent"
             android:layout_height="wrap_content"
+            android:layout_marginStart="20dp"
+            android:layout_marginEnd="20dp"
             android:background="@drawable/background_form_rounded"
             android:hint="@string/enter_password"
             android:inputType="textPassword"
@@ -98,6 +102,7 @@
         android:id="@+id/txtRememberMe"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
+        android:layout_marginStart="20dp"
         android:text="@string/remember_me"
         android:textColor="@color/white"
         android:textSize="14sp"
@@ -108,12 +113,13 @@
 
     <com.google.android.material.checkbox.MaterialCheckBox
         android:id="@+id/cbRememberMe"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
+        android:layout_width="48dp"
+        android:layout_height="48dp"
         android:layout_marginTop="8dp"
-        android:checked="true"
         android:scaleX="0.9"
         android:scaleY="0.9"
+        app:buttonCompat="@drawable/select_checkbox"
+        app:buttonTint="@color/white"
         app:layout_constraintStart_toEndOf="@id/txtRememberMe"
         app:layout_constraintTop_toBottomOf="@id/txtPasswordError" />
 
@@ -121,10 +127,12 @@
         android:id="@+id/btnLogin"
         android:layout_width="0dp"
         android:layout_height="wrap_content"
+        android:layout_marginStart="20dp"
         android:layout_marginTop="20dp"
-        android:background="@drawable/background_form_rounded"
+        android:layout_marginEnd="20dp"
+        android:background="@drawable/selector_auth_button_background"
         android:text="@string/login"
-        android:textColor="@color/black"
+        android:textColor="@color/auth_buttons_text_colors"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintTop_toBottomOf="@id/cbRememberMe" />
