+++ app/src/main/res/layout/fragment_register.xml
@@ -5,14 +5,13 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/blue"
-    android:padding="20dp"
     tools:context=".feature_register.presentation.screen.RegisterFragment">
 
     <androidx.appcompat.widget.AppCompatTextView
         android:id="@+id/txtRegisterHeader"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
-        android:layout_marginTop="2dp"
+        android:layout_marginTop="8dp"
         android:text="@string/register"
         android:textColor="@color/white"
         android:textSize="36sp"
@@ -25,8 +24,10 @@
         android:id="@+id/btnArrowBack"
         android:layout_width="24dp"
         android:layout_height="24dp"
+        android:layout_marginStart="20dp"
         android:background="@android:color/transparent"
         android:scaleType="centerCrop"
+
         android:src="@drawable/arrow_back"
         android:tint="@color/white"
         app:layout_constraintBottom_toBottomOf="@id/txtRegisterHeader"
@@ -37,11 +38,14 @@
         android:id="@+id/etEmail"
         android:layout_width="0dp"
         android:layout_height="wrap_content"
+        android:layout_marginStart="20dp"
         android:layout_marginTop="12dp"
+        android:layout_marginEnd="20dp"
         android:background="@drawable/background_form_rounded"
         android:drawableEnd="@drawable/ic_email"
         android:drawablePadding="12dp"
         android:hint="@string/enter_email"
+        android:inputType="textEmailAddress"
         android:maxLines="1"
         android:paddingStart="12dp"
         android:paddingTop="12dp"
@@ -82,6 +86,8 @@
             android:id="@+id/etPassword"
             android:layout_width="match_parent"
             android:layout_height="wrap_content"
+            android:layout_marginStart="20dp"
+            android:layout_marginEnd="20dp"
             android:background="@drawable/background_form_rounded"
             android:hint="@string/enter_password"
             android:inputType="textPassword"
@@ -121,9 +127,10 @@
 
         <com.google.android.material.textfield.TextInputEditText
             android:id="@+id/etRepeatPassword"
-
             android:layout_width="match_parent"
             android:layout_height="wrap_content"
+            android:layout_marginStart="20dp"
+            android:layout_marginEnd="20dp"
             android:background="@drawable/background_form_rounded"
             android:hint="@string/enter_password"
             android:inputType="textPassword"
@@ -153,10 +160,12 @@
         android:id="@+id/btnRegister"
         android:layout_width="0dp"
         android:layout_height="wrap_content"
+        android:layout_marginStart="20dp"
         android:layout_marginTop="20dp"
-        android:background="@drawable/background_form_rounded"
+        android:layout_marginEnd="20dp"
+        android:background="@drawable/selector_auth_button_background"
         android:text="@string/register"
-        android:textColor="@color/black"
+        android:textColor="@color/auth_buttons_text_colors"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintTop_toBottomOf="@id/txtRepeatPasswordError" />
