+++ app/src/main/res/values/colors.xml
@@ -7,4 +7,7 @@
     <color name="red">#CB292C</color>
     <color name="green">#33A937</color>
     <color name="gray">#8E8E8E</color>
+    <color name="transparent_blue">#37002737</color>
+    <color name="dark_white" >#CECECE</color>
+
 </resources>
